fix ads : edit crud operations(all)

+++ app/Http/Controllers/Api/User/AdController.php
@@ -2,59 +2,68 @@
 
 namespace App\Http\Controllers\Api\User;
 
-use App\Helpers\ApiResponse;
-use App\Http\Controllers\Controller;
-use App\Http\Requests\Adrequest;
-use App\Http\Resources\AdResource;
 use App\Models\Ad;
 use App\Services\AdService;
-use Illuminate\Contracts\Validation\Validator;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
+use App\Traits\ApiResponseTrait;
 use Spatie\FlareClient\Api;
+use Illuminate\Http\Request;
+use App\Http\Requests\Adrequest;
+use Illuminate\Http\JsonResponse;
+use App\Http\Resources\AdResource;
+use Illuminate\Support\Facades\DB;
+use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Contracts\Validation\Validator;
 
 class AdController extends Controller
 {
-    protected $adService;
+    use ApiResponseTrait;
 
-    public function __construct(AdService $adService)
+    protected static $adResourceclass = AdResource::class;
+
+    public function __construct(protected AdService $adService)
     {
-        $this->adService = $adService;
     }
 
-    public function index()
+    public function index(): mixed
     {
-        return $this->adService->index();
+        $ads = $this->adService->index();
+
+        return self::$adResourceclass::collection($ads);
     }
 
     public function store(Adrequest $request)
     {
-        $ad = $this->adService->storeAd($request);
-
-        if ($ad) return ApiResponse::success(201, 'Ad created', new AdResource($ad));
+        $this->adService->storeAd($request);
+        return $this->success([],__('site.Ad Added successfully'),200);
     }
 
-    public function show()
+    public function show(): mixed
     {
-        return $this->adService->show();
+        $ads= $this->adService->show();
+
+        return self::$adResourceclass::collection($ads);
+
     }
 
-    public function update(Adrequest $request, $adId)
+    public function update(Adrequest $request, $adId): JsonResponse
     {
-        //$this->adService->updateAd($request,  $adId);
-        $ad = Ad::findOrFail($adId);
-        if ($ad->user_id != auth()->user()->id)  return ApiResponse::error(403, 'You are not allwoed to do this action', []);
+        return $this->adService->updateAd($request, $adId) ?? null;
+      
 
-        $data = $request->validated();
-        $updatedAd = $ad->update($data);
-
-        if ($updatedAd) return ApiResponse::success(201, 'Updated successfully',new AdResource($ad));
     }
 
-    public function destroy($adId)
+    public function destroy($adId) :mixed
     {
+        $adIdExists = DB::table('ads')->where('id', $adId)->exists();
+
+        if (!$adIdExists) {
+
+            return $this->error( __('site.Ad not found'),404);
+        }
+
         $deletRecord = $this->adService->destroyAd($adId);
 
-        if ($deletRecord) return  ApiResponse::success(200, 'Deleted Successfully', []);
+        return $this->success([],__('site.Ad Deleted successfully'),200);
     }
 }
